refactor: type files
